refactor(routes/show): split components across files
This patch updates the show page definition to be split across multiple
files to more easily encapsulate concerns and avoid name clashes (e.g.
the `ExternalLink` icon v.s. the component).

See: https://remix.run/docs/en/main/file-conventions/route-files-v2#folders-for-organization
